Core/Misc: move back Liquid defines into MapDefines.h
