+++ src/common/Collision/Maps/MMapDefines.h
@@ -19,8 +19,6 @@
 #define MMapDefines_h__
 
 #include "Define.h"
-#include "MapDefines.h"
-#include "Optional.h"
 #include <DetourNavMesh.h>
 
 const uint32 MMAP_MAGIC = 0x4d4d4150; // 'MMAP'
@@ -69,49 +67,6 @@ enum NavTerrainFlag
     NAV_GROUND_STEEP = 1 << (NAV_AREA_MAX_VALUE - NAV_AREA_GROUND_STEEP),
     NAV_WATER        = 1 << (NAV_AREA_MAX_VALUE - NAV_AREA_WATER),
     NAV_MAGMA_SLIME  = 1 << (NAV_AREA_MAX_VALUE - NAV_AREA_MAGMA_SLIME)
-};
-
-enum ZLiquidStatus : uint32
-{
-    LIQUID_MAP_NO_WATER     = 0x00000000,
-    LIQUID_MAP_ABOVE_WATER  = 0x00000001,
-    LIQUID_MAP_WATER_WALK   = 0x00000002,
-    LIQUID_MAP_IN_WATER     = 0x00000004,
-    LIQUID_MAP_UNDER_WATER  = 0x00000008,
-};
-
-#define MAP_LIQUID_STATUS_SWIMMING (LIQUID_MAP_IN_WATER | LIQUID_MAP_UNDER_WATER)
-#define MAP_LIQUID_STATUS_IN_CONTACT (MAP_LIQUID_STATUS_SWIMMING | LIQUID_MAP_WATER_WALK)
-
-struct LiquidData
-{
-    EnumFlag<map_liquidHeaderTypeFlags> type_flags = map_liquidHeaderTypeFlags::NoWater;
-    uint32 entry;
-    float  level;
-    float  depth_level;
-};
-
-struct WmoLocation
-{
-    WmoLocation() = default;
-    WmoLocation(int32 groupId, int32 nameSetId, int32 rootId, uint32 uniqueId)
-        : GroupId(groupId), NameSetId(nameSetId), RootId(rootId), UniqueId(uniqueId) { }
-
-    int32 GroupId = 0;
-    int32 NameSetId = 0;
-    int32 RootId = 0;
-    uint32 UniqueId = 0;
-};
-
-struct PositionFullTerrainStatus
-{
-    PositionFullTerrainStatus() : areaId(0), floorZ(0.0f), outdoors(true), liquidStatus(LIQUID_MAP_NO_WATER) { }
-    uint32 areaId;
-    float floorZ;
-    bool outdoors;
-    ZLiquidStatus liquidStatus;
-    Optional<WmoLocation> wmoLocation;
-    Optional<LiquidData> liquidInfo;
 };
 
 #endif // MMapDefines_h__
